improved button functionally, and divided screens into individual files
