+++ main.py
@@ -1,0 +1,7 @@
+from screens.splash_screen import *
+from screens.player_screen import *
+from screens.action_screen import *
+
+splash_screen()
+player_screen()
+action_screen()
